--- dags/ethereum_parse_v2_datasets.py
@@ -1,1 +1,0 @@
-PARSE_V2_DATASETS = {'aelf', 'curve'}
